# docs/swagger.yaml
basePath: /
definitions:
  handlers.RefreshRequest:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  tokens.TokenPair:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Документация API для сервиса аутентификации
  title: Auth Service API
  version: "1.0"
paths:
...
  /tokens/refresh:
    post:
      consumes:
      - application/json
      description: Обновляет access и refresh токены при валидной паре
      parameters:
      - description: Refresh tokens
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tokens.TokenPair'
        "400":
          description: invalid request
          schema:
            type: string
        "401":
          description: unauthorized
          schema:
            type: string
      summary: Обновление токенов
      tags:
      - auth
swagger: "2.0"